Move files to correct file and delete unused file
